Move PropertyDeclarationBlock into its own module.
